feat(bom): allow decimal values for quantity_per_unit
- Updated validation in BomItemController to accept numeric instead of integer for quantity_per_unit in store and update methods
- Removed integer cast for quantity_per_unit in BomItem model to support decimals
- Changed database migration to use decimal type for quantity_per_unit column

This enables fractional quantities in Bill of Materials items for more accurate material requirements.

+++ database/migrations/2025_12_18_183314_create_bom_items_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
                 ->on('materials')
                 ->onDelete('restrict');
 
-            $table->integer('quantity_per_unit');
+            $table->decimal('quantity_per_unit');
             $table->integer('total_required');
 
             $table->integer('allocated');
